+++ src/WebJobs.Script/Workers/Rpc/IWorkerConfigurationResolver.cs
@@ -8,8 +8,8 @@ namespace Microsoft.Azure.WebJobs.Script.Workers.Rpc.Configuration
     /// <summary>
     /// Interface to resolve the Worker Configs.
     /// </summary>
-    public interface IWorkerConfigurationResolver
+    internal interface IWorkerConfigurationResolver
     {
-        List<string> GetWorkerConfigs();
+        List<string> GetWorkerConfigPaths();
     }
 }