+++ test/WebJobs.Script.Tests/Workers/Rpc/WorkerConfigurationResolverTests.cs
@@ -35,6 +35,9 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Workers.Rpc
         [InlineData("LATEST", "java\\2.19.0", "node\\3.10.1", "powershell\\7.4", "dotnet-isolated", "python")]
         [InlineData("STANDARD", "java\\2.18.0", "node\\3.10.1", "powershell\\7.4", "dotnet-isolated", "python")]
         [InlineData("EXTENDED", "java\\2.18.0", "node\\3.10.1", "powershell\\7.4", "dotnet-isolated", "python")]
+        [InlineData("laTest", "java\\2.19.0", "node\\3.10.1", "powershell\\7.4", "dotnet-isolated", "python")]
+        [InlineData("abc", "java\\2.19.0", "node\\3.10.1", "powershell\\7.4", "dotnet-isolated", "python")]
+        [InlineData("Standard", "java\\2.18.0", "node\\3.10.1", "powershell\\7.4", "dotnet-isolated", "python")]
         public void GetWorkerConfigs_MultiLanguageWorker_ReturnsExpectedConfigs(string releaseChannel, string java, string node, string powershell, string dotnetIsolated, string python)
         {
             // Arrange
@@ -45,7 +48,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Workers.Rpc
 
             var workerConfigurationResolver = new DynamicWorkerConfigurationResolver(_mockConfig.Object, _mockLogger.Object, mockEnvironment.Object, FileUtility.Instance, _mockProfileManager.Object, new HashSet<string>() { "java", "node", "powershell" }, _probingPaths);
 
-            var result = workerConfigurationResolver.GetWorkerConfigs();
+            var result = workerConfigurationResolver.GetWorkerConfigPaths();
 
             // Assert
             Assert.Equal(result.Count, 5);
@@ -60,7 +63,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Workers.Rpc
         [InlineData(null, "LATEST")]
         [InlineData(null, "STANDARD")]
         [InlineData("Empty", "LATEST")]
-        [InlineData("Empty", "STANDARD")]
+        [InlineData("Empty", "abc")]
         public void GetWorkerConfigs_MultiLanguageWorker_NullOREmptyProbingPath_ReturnsExpectedConfigs(string probingPathValue, string releaseChannel)
         {
             // Arrange
@@ -79,7 +82,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Workers.Rpc
             // Act
             var workerConfigurationResolver = new DynamicWorkerConfigurationResolver(_mockConfig.Object, _mockLogger.Object, mockEnvironment.Object, FileUtility.Instance, _mockProfileManager.Object, new HashSet<string>() { "java", "node", "powershell" }, probingPaths);
 
-            var result = workerConfigurationResolver.GetWorkerConfigs();
+            var result = workerConfigurationResolver.GetWorkerConfigPaths();
 
             // Assert
             Assert.Equal(result.Count, 5);
@@ -124,7 +127,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Workers.Rpc
             // Act
             var workerConfigurationResolver = new DynamicWorkerConfigurationResolver(mockConfig.Object, mockLogger.Object, mockEnv.Object, FileUtility.Instance, mockProfileManager.Object, new HashSet<string>() { "java", "node", "powershell" }, probingPaths);
 
-            var result = workerConfigurationResolver.GetWorkerConfigs();
+            var result = workerConfigurationResolver.GetWorkerConfigPaths();
 
             // Assert
             Assert.Equal(result.Count, 1);